requirements.txt
asgiref==3.7.2
Django==5.0.3
djangorestframework==3.15.0
pillow==10.2.0
psycopg2==2.9.9
python-dotenv==1.0.0
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
